Change types to size_t where appropriate and fix libraries

+++ include/blur.h
@@ -7,8 +7,9 @@
 #ifndef BLUR_H
 #define BLUR_H
 
-#include <stdint.h>
-#include <math.h>
+#include <cstdint>
+#include <cstdlib>
+#include <cmath>
 
 /**************************************************************************//**
  * Box blur main function.
@@ -20,7 +21,6 @@
  * @param[in] s Source image
  * @param[out] d Destination image
  *****************************************************************************/
-void blur(uint32_t w, uint32_t h, uint32_t r, uint32_t wt, float_t* s,
-    float_t* d);
+void blur(size_t w, size_t h, uint32_t r, uint32_t wt, float_t* s, float_t* d);
 
 #endif
